# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NaissanceChain API Documentation",
        "description": "API complète pour la gestion des actes de naissance avec blockchain et QR codes",
        "version": "1.0.0",
        "contact": {
            "email": "support@naissancechain.gn",
            "name": "Équipe NaissanceChain"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API de Développement"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer Token"
            }
        },
        "schemas": {
            "Naissance": {
                "type": "object",
                "title": "Acte de Naissance",
                "description": "Modèle représentant un acte de naissance dans NaissanceChain",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID unique de l'acte",
                        "example": 1
                    },
                    "numero_unique": {
                        "type": "string",
                        "description": "Numéro unique de l'acte généré par le système",
                        "example": "NC-2026-000001"
                    },
                    "nom_enfant": {
                        "type": "string",
                        "description": "Nom de famille de l'enfant",
                        "example": "TOURE"
                    },
                    "prenom_enfant": {
                        "type": "string",
                        "description": "Prénom de l'enfant",
                        "example": "Ibrahim"
                    },
                    "date_naissance": {
                        "type": "string",
                        "format": "date",
                        "description": "Date de naissance de l'enfant",
                        "example": "2026-03-20"
                    },
                    "lieu_naissance": {
                        "type": "string",
                        "description": "Lieu de naissance",
                        "example": "Hôpital Donka, Conakry"
                    },
                    "sexe": {
                        "type": "string",
                        "enum": ["M", "F"],
                        "description": "Sexe de l'enfant",
                        "example": "M"
                    },
                    "numero_acte": {
                        "type": "string",
                        "description": "Numéro officiel de l'acte",
                        "example": "ACTE-2026-002"
                    },
                    "statut": {
                        "type": "string",
                        "enum": ["valide", "en_attente", "synchronise", "annule"],
                        "description": "Statut de l'acte",
                        "example": "valide"
                    },
                    "hash_sha256": {
                        "type": "string",
                        "description": "Hash SHA256 pour l'intégrité blockchain",
                        "example": "abc123def456789..."
                    },
                    "qr_code_url": {
                        "type": "string",
                        "description": "URL du code QR généré",
                        "example": "/storage/qr/abc123.png"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Date de création de l'enregistrement",
                        "example": "2026-04-30T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Date de dernière modification",
                        "example": "2026-04-30T10:00:00Z"
                    }
                }
            }
        }
    },
    "paths": {
...
        "/naissances": {
            "post": {
                "tags": ["Naissances"],
                "summary": "Créer un nouvel acte de naissance",
                "description": "Crée un nouvel acte de naissance avec génération QR code et stockage blockchain",
                "operationId": "createNaissance",
                "security": [{"BearerAuth": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["nom_enfant", "prenom_enfant", "date_naissance", "lieu_naissance", "sexe", "nom_pere", "prenom_pere", "nom_mere", "prenom_mere", "declarant_nom", "declarant_prenom", "declarant_lien", "officier_etat_civil", "numero_acte", "date_enregistrement"],
                                "properties": {
                                    "nom_enfant": {"type": "string", "example": "TOURE", "description": "Nom de l'enfant"},
                                    "prenom_enfant": {"type": "string", "example": "Ibrahim", "description": "Prénom de l'enfant"},
                                    "date_naissance": {"type": "string", "format": "date", "example": "2026-03-20", "description": "Date de naissance"},
                                    "lieu_naissance": {"type": "string", "example": "Hôpital Donka, Conakry", "description": "Lieu de naissance"},
                                    "sexe": {"type": "string", "enum": ["M", "F"], "example": "M", "description": "Sexe de l'enfant"},
                                    "numero_acte": {"type": "string", "example": "ACTE-2026-002", "description": "Numéro de l'acte"},
                                    "date_enregistrement": {"type": "string", "format": "date-time", "example": "2026-04-30T10:00:00Z", "description": "Date d'enregistrement"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Acte de naissance créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {"type": "boolean", "example": true},
                                        "message": {"type": "string", "example": "Acte de naissance créé avec succès"},
                                        "data": {
                                            "$ref": "#/components/schemas/Naissance"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}